[#486] meetup view sorted
